requirements.txt
autopep8==1.4.4
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycodestyle==2.5.0
pymongo==3.9.0
python-dotenv==0.10.3
Werkzeug==0.16.0
